Separated CategoryDao
