+++ dao/src/main/java/by/it/academy/dao/NewsDAO.java
@@ -166,83 +166,6 @@ public class NewsDAO implements DAO {
 		return list;
 	}
 
-	public List<Category> getAllCategories() {
-        Connection connectiona = null;
-        try {
-            connectiona = DataSource.getInstance().getConnection();
-        } catch (SQLException e) {
-            e.printStackTrace();
-        }
-        List<Category> list = new ArrayList<Category>();
-		String query = "SELECT * FROM category";
-		try {
-			Statement st = connectiona.createStatement();
-			ResultSet result = st.executeQuery(query);
-			while (result.next()) {
-				Category category = new Category();
-				category.setCatId(result.getString(1));
-				category.setParentId(result.getString(2));
-				list.add(category);
-			}
-			result.close();
-			st.close();
-		} catch (SQLException e) {
-			logger.error("Error get all categories", e);
-		}
-        try {
-            connectiona.close();
-        } catch (SQLException e) {
-            logger.error(e);
-        }
-		return list;
-	}
-
-	public Category getCategoryById(String id) {
-		Category cat = null;
-		String query = "SELECT * FROM category WHERE catId=?";
-		try {
-			PreparedStatement pStatement = connection.prepareStatement(query);
-			pStatement.setString(1, id);
-			ResultSet result = pStatement.executeQuery();
-
-			if (result.next()) {
-				cat = new Category();
-				cat.setCatId(result.getString(1));
-				cat.setParentId(result.getString(2));
-			}
-
-		} catch (SQLException e) {
-			logger.error("Error get category by id", e);
-		}
-        try {
-            connection.close();
-        } catch (SQLException e) {
-            logger.error(e);
-        }
-		return cat;
-	}
-
-	public int addCat(Category category) {
-		String query = "INSERT INTO category VALUES (?,?)";
-		int result = 0;
-		PreparedStatement pStatement;
-		try {
-			pStatement = connection.prepareStatement(query);
-			pStatement.setString(1, category.getCatId());
-			pStatement.setString(2, category.getParentId());
-			result = pStatement.executeUpdate();
-
-		} catch (SQLException e) {
-			logger.error("Error get add category", e);
-		}
-        try {
-            connection.close();
-        } catch (SQLException e) {
-            logger.error(e);
-        }
-		return result;
-	}
-
 	public List<News> getNewsByCategoryId(String category) {
 		List<News> list = new ArrayList<News>();
 		String query = "SELECT * FROM news WHERE categoryId=" + "'" + category + "'";
